fix: compilation errors — VmFlags, SocketHandle, type mismatches, missing functions
- Replace bitflags! macro with manual VmFlags impl (no bitflags dependency)
- Add BitOrAssign for VmFlags (fixes |= operator)
- Extract socket_handle_id() helper for smoltcp SocketHandle (private field)
- Fix io_uring dispatch wrapper argument types (u64→u32 casts)
- Split combined EventFd|TimerFd|IoUringFd match arm (different inner types)
- Add yield_now() to scheduler module (used by mqueue blocking)
- Fix check_signal_interrupt import path (syscalls:: not task::process::)
- Fix sa_restorer out-of-scope (use restorer pointer comparison instead)
- Fix switch.rs brace mismatch (extra closing brace)
- Add fpu_state/sched_class to Thread initializers in test files
- Fix RTT estimation type mismatch (u32→u64 casts in tcp_congestion)

+++ kernel/src/net/tcp_congestion.rs
@@ -208,9 +208,9 @@ fn update_rtt(s: &mut TcpCongestionState, rtt_ms: u64) {
             s.srtt_us - rtt_us
         };
         // RTTVAR = (3/4) * RTTVAR + (1/4) * |diff|
-        s.rttvar_us = ((BETA - 1) * s.rttvar_us + diff) / BETA;
+        s.rttvar_us = (((BETA - 1) as u64) * s.rttvar_us + diff) / (BETA as u64);
         // SRTT = (7/8) * SRTT + (1/8) * R
-        s.srtt_us = ((ALPHA - 1) * s.srtt_us + rtt_us) / ALPHA;
+        s.srtt_us = (((ALPHA - 1) as u64) * s.srtt_us + rtt_us) / (ALPHA as u64);
     }
     // RTO = SRTT + 4 * RTTVAR, clamped to [200ms, 60s].
     let rto_us = s.srtt_us.saturating_add(4 * s.rttvar_us);
